autodesk-agents-unified/requirements.txt
# Core FastAPI and server dependencies
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0

# Configuration and environment
pyyaml==6.0.1
python-dotenv>=1.0.0

# AWS and Bedrock integration
boto3>=1.34.72
langchain==0.1.0
langchain-aws==0.1.0
langchain-community==0.0.10

# Vector search and database
opensearch-py==2.4.0
requests-aws4auth>=1.1.2

# Authentication and HTTP
aiohttp>=3.9.0
pyjwt==2.8.0
requests>=2.31.0

# File and async operations
aiofiles==23.2.1

# System monitoring
psutil==5.9.6

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Utilities
jinja2>=3.1.0
markdown>=3.5.0